requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pillow==10.1.0
python-magic==0.4.27
openai==1.3.7
googletrans==4.0.0rc1
speechrecognition==3.10.0
pydub==0.25.1
gtts==2.4.0
pytesseract==0.3.10
nltk==3.8.1
textblob==0.17.1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
